:recycle: (Venue) Extract part of venuebody to put in venuecontent
