+++ src/features/venue/components/VenueBody/VenueBody.tsx
@@ -1,7 +1,6 @@
 import React, { FunctionComponent } from 'react'
 import { View } from 'react-native'
-import { IOScrollView as IntersectionObserverScrollView } from 'react-native-intersection-observer'
-import styled, { useTheme } from 'styled-components/native'
+import { useTheme } from 'styled-components/native'
 
 import { VenueResponse } from 'api/gen'
 import { GtlPlaylistData } from 'features/gtlPlaylist/types'
@@ -12,15 +11,12 @@ import { VENUE_CTA_HEIGHT_IN_SPACES } from 'features/venue/components/VenueCTA/V
 import { VenueMessagingApps } from 'features/venue/components/VenueMessagingApps/VenueMessagingApps'
 import { VenueOffers } from 'features/venue/components/VenueOffers/VenueOffers'
 import { VenueThematicSection } from 'features/venue/components/VenueThematicSection/VenueThematicSection'
-import { VenueTopComponent } from 'features/venue/components/VenueTopComponent/VenueTopComponent'
 import { analytics } from 'libs/analytics'
-import { useGetHeaderHeight } from 'ui/components/headers/PageHeaderWithoutPlaceholder'
 import { SectionWithDivider } from 'ui/components/SectionWithDivider'
 import { Spacer } from 'ui/theme'
 
 interface Props {
   venue: VenueResponse
-  onScroll: () => void
   venueOffers?: VenueOffersType
   playlists?: GtlPlaylistData[]
   shouldDisplayCTA?: boolean
@@ -28,24 +24,17 @@ interface Props {
 
 export const VenueBody: FunctionComponent<Props> = ({
   venue,
-  onScroll,
   venueOffers,
   playlists,
   shouldDisplayCTA,
 }) => {
   const { isDesktopViewport, isTabletViewport } = useTheme()
-  const headerHeight = useGetHeaderHeight()
   const isLargeScreen = isDesktopViewport || isTabletViewport
 
   const FirstSectionContainer = isLargeScreen ? View : SectionWithDivider
 
   return (
-    <Container onScroll={onScroll} scrollEventThrottle={16} bounces={false}>
-      {isLargeScreen ? <Placeholder height={headerHeight} /> : null}
-      <VenueTopComponent venue={venue} />
-
-      <Spacer.Column numberOfSpaces={isDesktopViewport ? 10 : 6} />
-
+    <React.Fragment>
       <FirstSectionContainer visible gap={6}>
         <TabLayout
           tabPanels={{
@@ -79,16 +68,6 @@ export const VenueBody: FunctionComponent<Props> = ({
       <SectionWithDivider visible={!!shouldDisplayCTA} gap={VENUE_CTA_HEIGHT_IN_SPACES}>
         <Spacer.Column numberOfSpaces={6} />
       </SectionWithDivider>
-    </Container>
+    </React.Fragment>
   )
 }
-
-const Container = styled(IntersectionObserverScrollView).attrs({
-  scrollIndicatorInsets: { right: 1 },
-})({
-  overflow: 'visible',
-})
-
-const Placeholder = styled.View<{ height: number }>(({ height }) => ({
-  height,
-}))
